% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{addRowBtn}
\alias{addRowBtn}
\title{Create button to add a row}
\usage{
addRowBtn(tabinput)
}
\arguments{
\item{tabinput}{List containing specifications for table inputs.}
}
\description{
Create button to add a row
}
\keyword{internal}
