% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{customSelectInput}
\alias{customSelectInput}
\title{Create custom select input}
\usage{
customSelectInput(input, options = NULL)
}
\arguments{
\item{input}{List containing input config.}

\item{options}{Vector of select options. If not given, will be extracted from input specs.}
}
\description{
Create custom select input
}
