% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draftStandardDSP}
\alias{draftStandardDSP}
\title{Draft DSP with recommended bundle}
\usage{
draftStandardDSP(dir, verbose = TRUE)
}
\arguments{
\item{dir}{Directory folder to create; forms will be organized here and it will be set as working directory.}

\item{verbose}{Whether to be verbose.}
}
\description{
This is a convenience function to bring all the "standard"
paperwork templates into the draft space.
It is updated to reflect the latest concept of what is standard,
i.e. some sections might be added or removed from this bundle over time.
}
