% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_processing.R
\name{formChecklistContext}
\alias{formChecklistContext}
\title{Contextual handler for creating a form checklist}
\usage{
formChecklistContext(
  config_file,
  mode = knitr::opts_knit$get("rmarkdown.pandoc.to")
)
}
\arguments{
\item{config_file}{Path to YAML config file}

\item{mode}{The render mode.}
}
\description{
This only make sense for HTML version of the form.
}
