% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_processing.R
\name{formInputContext}
\alias{formInputContext}
\title{Contextual handler for rendering input as HTML form or PDF doc}
\usage{
formInputContext(
  config_file = NULL,
  data = NULL,
  mode = knitr::opts_knit$get("rmarkdown.pandoc.to")
)
}
\arguments{
\item{config_file}{Path to YAML config file}

\item{data}{Path to JSON data.}

\item{mode}{The render mode.}
}
\description{
Given \code{mode} (which is expected to be set via \code{knitr} runtime parameters
so it doesn't need to be passed in), this takes care of logic for
reading in the correct input dependency and any additional processing for said input.
If \code{mode} == "html", we expect to read a \code{config_file}.
If \code{mode} == "latex", we expect to read JSON data and process it to correct format.
}
