% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{formSubmit}
\alias{formSubmit}
\title{Element to submit form}
\usage{
formSubmit(config_submit)
}
\arguments{
\item{config_submit}{List containing config for submit element.}
}
\description{
Currently this is just a shorthand that calls\code{formSubmitButton} because
the only configuration allowed is a button.
}
