% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{formSubmitButton}
\alias{formSubmitButton}
\title{Create button to submit form}
\usage{
formSubmitButton(id, action = "https://submit-form.com/echo", label = "Submit")
}
\arguments{
\item{id}{Id of the form.}

\item{action}{Form action URL (defaults to a test URL).}

\item{label}{Button label.}
}
\description{
Create button to submit form
}
