% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_processing.R
\name{fromRadioButtonInput}
\alias{fromRadioButtonInput}
\title{Reconstruct radio button inputs}
\usage{
fromRadioButtonInput(.F)
}
\arguments{
\item{.F}{Form data object.}
}
\description{
Radio button inputs have two inputs for "Yes" and "No",
e.g. Input_1_Y and Input_1_N, both tied to Input_1.
Since the data returned will be Input_1 and either "Y"or "No" as the value,
we reconstruct "Input_1_*" by taking the value present in Input_1.
If it's "Y", "Input_1_Y" will be constructed, or "Input_1_N" if otherwise.
Currently the input specs have to make sure that input names match the
value option for reconstruction to work.
This is also currently hard-coded for Addendum inputs but should be generalized
for all radio button inputs (relying on the config to do so).
}
