% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_processing.R
\name{fromTableInput}
\alias{fromTableInput}
\title{Reconstruct table inputs}
\usage{
fromTableInput(.F)
}
\arguments{
\item{.F}{Form data object.}
}
\description{
Inputs from table needs to be set into the parent as a \code{data.frame}.
}
