% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getConfig}
\alias{getConfig}
\title{Create config object in global env}
\usage{
getConfig(yaml)
}
\arguments{
\item{yaml}{YAML configuration file.}
}
\description{
The config object needs to be available to provide parameters to form input functions.
This first calls upon \code{readConfig} to read the YAML file.
However, depending on whether a global object \code{.F} aready exists,
it will append the new config.
}
\details{
TO DO: Warn if input namespaces confict.
TO DO: Better way to handle/check .F global object?
}
