% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{newCell}
\alias{newCell}
\title{Wrap content in a table cell}
\usage{
newCell(content, header = FALSE)
}
\arguments{
\item{content}{Content that becomes inner HTML of the cell.}

\item{header}{Whether this cell is part of a table header, default \code{FALSE}.}
}
\description{
Wrap content in a table cell
}
