% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{newChecklist}
\alias{newChecklist}
\title{Create a checklist}
\usage{
newChecklist(config)
}
\arguments{
\item{config}{List with checklist configuration.}
}
\description{
Wrapper to create a checklist element according to given configuration.
}
