% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{newInput}
\alias{newInput}
\title{Create HTML form input}
\usage{
newInput(input)
}
\arguments{
\item{input}{List containing input config.}
}
\description{
Handles common types \code{c("select", "text", "textarea", "date", "number", "email")}
as well as specially defined types called \code{table-inputs} and \code{select-schema}.
}
