% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{newRow}
\alias{newRow}
\title{Wrap cells in table row}
\usage{
newRow(cells, removable = TRUE)
}
\arguments{
\item{cells}{A list or taglist of table cell elements.}

\item{removable}{Whether this row is removable, which appends a button with
the functionality.}
}
\description{
Wrap cells in table row
}
