% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{RIA}
\alias{RIA}
\title{RIA: Render Input Appropriately}
\usage{
RIA(x, mode = knitr::opts_knit$get("rmarkdown.pandoc.to"))
}
\arguments{
\item{x}{Either an input specification (for HTML) or data value to render (for PDF).}

\item{mode}{(Optional) Render context, which can be extracted at run from variable set by \code{knitr}.}
}
\description{
RIA: Render Input Appropriately
}
