% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{schemaRange}
\alias{schemaRange}
\title{Get classes from a JSON-LD schema}
\usage{
schemaRange(jsonld, id, warn = FALSE)
}
\arguments{
\item{jsonld}{URL or path to local json-ld file from which schema will be read.}

\item{id}{Id (i.e. in \verb{@id}) for which to get range values.}

\item{warn}{Whether to warn if there are issues in the schema.}
}
\description{
Note: Assumes schematic-generated json-ld file.
This looks up defined valid values (range) given the \verb{@id} as encoded in schema.
}
