% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{schemaSelectInput}
\alias{schemaSelectInput}
\title{Create select input with controlled values from schema}
\usage{
schemaSelectInput(input)
}
\arguments{
\item{input}{List containing input config.}
}
\description{
Create a select input where options are schema-based,
i.e. defined by a specified class such as "Assay"
}
