% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{tableInputWrapper}
\alias{tableInputWrapper}
\title{Wrapper to render table input}
\usage{
tableInputWrapper(tabinput)
}
\arguments{
\item{tabinput}{List containing specifications for table inputs.}
}
\description{
Wrapper to render table input
}
