% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{tableWithInput}
\alias{tableWithInput}
\title{Create form elements with table layout}
\usage{
tableWithInput(tabinput, nrow = 3, removable = TRUE)
}
\arguments{
\item{tabinput}{List containing specifications for table inputs.}

\item{nrow}{Number of rows, i.e. how many times to repeat.}

\item{removable}{Whether rows can be deleted, by default TRUE.}
}
\description{
Sometimes it is highly preferable to organize repeated elements in a table,
similar to a banking form where one might have multiple rows each with columns:
bank account name, bank account type, bank account number.
}
