% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draftStandardDSP}
\alias{draftStandardDSP}
\title{Draft DSP with recommended bundle}
\usage{
draftStandardDSP(version = c("v2", "v1"), verbose = TRUE)
}
\arguments{
\item{version}{Desired version of document; defaults to latest, which is v2.}

\item{verbose}{Whether to be verbose.}
}
\description{
Bring all the "standard" template components into the draft space.
This should reflect the latest concept of what is standard,
i.e. some sections might be added or removed from this bundle over time.
}
